SWANDROID-3018: AVKO doesn't start polling for CEC while RX SENSE is false
Fixed issue where the STB wouldn't begin CEC polling if the RxSense
is low.   This can cause the STB to incorrectly allocate a logical
address if an HDMI hotplug occurs with a TV that is in standby
(and RxSense is low).

For example, if the user disconnects the HDMI cable from say HDMI-1
on the TV and re-inserts it in to HDMI-2 whilst the TV is in standby,
then the CEC logical address is not re-determined.

The fix is as simple as allowing the HDMI hotplug "connected" event
to be passed to the Android framework even if the RxSense is low.

refs #SWANDROID-3018

Change-Id: Ib2227e08e34ef3512c340bd257f553046c514b95