SWANDROID-2848: 7252SSFFG can't shutdown then startup
Added support for placing Nexus in to S3 deep sleep when shutting
down Android.  This prevents the system from entering a half-way
house state whereby the kernel and Android have powered off but
Nexus has not.

With these changes, it is now possible to restart the STB using
any one of the usual wakeup methods (e.g. IR, BT, CEC).  The STB
will no longer unintentionally awaken on the spurious keypad
event but can still wake-up on timer/alarm events.  There will
need to be kernel changes in order to disable the waketimer from
waking up the system on alarm events.

A new monitor thread has been created in the Power HAL to check
for when the system is about to shutdown.  The shutdown event is
dispatched from the droid_pm PM kernel driver and the Power HAL
can then use this to place Nexus in to S3 deep sleep before
the system is then allowed to fully shutdown.

As the various sub-systems of Android do not all react to the
shutdown Intent, the NxClient standby monitoring code is used to
force each module to the standby state.  This can be seen in the
HWC and Audio HAL modules.

One additional change is required in refsw to remove the Android
reboot/shutdown notifier from the Nexus platform kernel driver code.

refs #SWANDROID-2848

Change-Id: Ifbc4d9549e8a2b99a56cd836c7d61c7956b4f2c8