hwc: display retire fence configurability.
default settings value: do not provide retire fence.

retire fence are used for the animation loop decisions in surface flinger;
a problem is that we cannot guarantee to provide the accurate data at the
real time display rate since it depends on the underlying nexus surface compositor
retire buffer callback scheme which may not actually trigger on each buffer swap
(as the queue is triple buffered and per nexus design); this in turn leads to
potentially providing inacurate fence signal rate back to the surface flinger
which translates into inacurate/delayed decisions and actions.

by disabling the retire fence we let surface flinger fall back to the prior
decision scheme based on vsync report instead; this addresses some of the shortfall
of the innacurate fence based scheme.  as we re-enable the retire fence in the future,
we also need to take a look at tuning the vsync event phase offset(s) which we have set
to 0 (default) right now but which will need to be more accurately tuned.

note: this does not remove the v3d fences which are used by the gl-consumer for
all v3d bound operations on buffers.

refs #SWANDROID-2995

Change-Id: I3eecd337348f9fee4a17ae52d5117c9af943fd6c
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>