hwc: bypass skip composition layer if we *must* restore background.
this situation arises during certain transition in particular from home
screen which has an underneath active video layer back into the foreground
video application (e.g. seen with youtube).

in this case, the skip composition optimization will result in purposefully
aborting a composition cycle, consequently the needed transparent background
would also be skipped and this would lead to a quick black screen flash in
front of the video layer.

to address this, we simply undo the skipping cycle for this composition if
we know we must restore the transparent background resulting in a smooth
transition.

refs #SWANDROID-2878

Change-Id: Ibdec71cecf11d668a5d5ab37f2188d2aaa2fda3a
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>