SWANDROID-3051: [PM] Cannot wakeup after system resume/suspend from a timer
The main problem is that when the system suspends to S2 (or S3) and resumes
from an alarm timer, a subsequent suspend fails to correctly place the full
system in to S2/S3 standby.  It is then not possible to wake up the system
without doing a full reboot.

The reason for this problem is because when an alarm timer expires causing
the kernel to resume from the suspended state, Nexus remains in standby and
does not inform security blocks to resume.  The security blocks need to be
made aware of each standby transition, otherwise the system will not fully
enter standby (and therefore cannot exit it either).

The changes now allow Nexus to resume from standby (i.e. enter S0) when an
alarm timer expires whilst in the suspended state.  When the system
subsequently suspends again, Nexus performs the same steps as before to
correctly place the system in to S2/S3 standby.  The nexus_ipc_client
"getPowerState()" API has been changed to "getPowerStatus()" to allow the
wakeup status to be returned to clients.  This is necessary to determine
whether the system resumed from an alarm timer or not.

These changes are also dependent on SWSTB-1292 to allow nexus to switch
quickly between S2, S0 and S2.   The polling for the standby transition
now uses the new nxclient standby state information to avoid losing the
double standby state transition. Proper synchronisation has been added
to NexusNxService to avoid deadlocking the system by unintentionally
making non-standby related Nexus/NxClient API calls whilst in standby.

refs #SWANDROID-3051

Change-Id: Ie1ebef025436c30bd6f42030babc6c5c2d0757e4