BOMX: Improve display path
Make the display-frame binder call asynchronous
in order to avoid blocking the binder while waiting
for a frame to be displayed. Increase the priority
of the component thread in order to serve the display
request promptly.
Also, added some code to print out the dropped
frames summary at the end of playback.

refs #SWANDROID-3098

Change-Id: Ia9397afe49632c8921cb35e6baafba11334eb0cc
Signed-off-by: Luis Navarro <navarro@broadcom.com>