SWANDROID-3111: [PM] Do not set GPIO's if partially coming out of standby
These changes are required to prevent Android from unintentionally
being woken up to the desktop/launcher when placed in to standby
with WoBLE enabled and an Alarm Timer fires.

The changes simply ensure that if the system is partially woken up,
then the state of any GPIO outputs is not changed.   This ensures
that the BT_DEV_WAKE signal is not momentarily toggled whilst
the system wakes up due to an alarm timer event and suspends again.

Without these changes, when an alarm timer fires whilst in standby,
the BT_DEV_WAKE signal will toggle causing the BT firmware to also
toggle the BT_HOST_WAKE signal.  The main SoC will detect this as
a valid WoBLE event causing a POWER key event to be spoofed.  This
will then cause Android to exit the sleep state and for the desktop
launcher to appear.

refs #SWANDROID-3111

Change-Id: I4a21d288256feaeeff3a6f4265ddb83a6b1f1c52