SWANDROID-3112: [PM] PmLibService fails to set state when coming out of standby
This set of changes allows for the original S0 PmLib state to be
retrieved and used in future "setState()" calls to help determine
the optimal PmLibsettings.  This is necessary to avoid the case of
trying to toggle a setting that ought not to be changed
(e.g. trying to enable SATA when SATA is physically disabled).

With this set of changes it is now possible to transition to S0
without generating an error.  The result of this is that a
subsequent transition to S0.5/S1 standby will result in all the
settings being acted upon (e.g. CPU1 taken off-line).

refs #SWANDROID-3112

Change-Id: I9cf26084656ca4bd727adceffb7994449bba5de7