nxservice: expose the 'display settings changed' callback.
this is a new event listener hook up allowing clients to register
for display settings change callback information.  this event is
distinct from hotplug callback since it can be trigger due to
internal only actions resulting in the display configuration change.

as example when an application switches dynamically output resolution
based on user input while connected to an output, such callback will
fire (but not the hotplug callback).

agents interested in receiving this information need to implement
the interface binder and register for such trigger.

refs #SWANDROID-2808

Change-Id: I585ed49d2727a246eb5ac6ccc2348720ac8be0c2