SWANDROID-3116: [PM] Cannot disable and re-enable Ethernet interface(s) with PmLibService
The change from Lollipop to Marshmallow has resulted in the "netcfg"
utility being removed and "ifconfig" is now used for low-level
control of the network interfaces.  As a result of this change,
PmLibService has now been updated to make use of this utility
instead.  The code has been tied up to use constant string
definitions instead of hardcoded string values.

Tested with S1 standby and the "eth0" network interface is taken
down and then back up successfully when resuming from S1 standby.

refs #SWANDROID-3116

Change-Id: Ia7a864382af161c46bd420f0c13253f830a21028