SWANDROID-3115: [PM] Provide additional checking for invalid entries in the aon_gpio config file
If an entry for a GPIO/SGPIO/AON_GPIO pin is invalid, then it won't increment
a counter.  If the number of entries in the aon_gpio.cfg file does not match
the counter at the end of the parsing process, then one or more entries is
invalid and this will be reported back as an error to the user through logcat.

Tested on a BCM97252SSFFG platform with a modified aon_gpio.cfg file containing
an invalid entry for "AON_GPIO65" (rather than "GPIO65").  This error was
logged successfully.

refs #SWANDROID-3115

Change-Id: I84e6eebd29ace0f83e5c546bf117c737105f1fde