SWANDROID-3091: [PM] Align using Android-M HDMI hotplug wakeup property
As the Android framework will not wake-up on an HDMI hot-plug
"connected" event if the Android-M property "ro.hdmi.wake_on_hotplug"
is set to "false" (which it is in our case), then it is not possible
to use this as a valid means for waking up the STB on receipt of
one of the few CEC messages whilst in standby.

In order to wake the STB up now on a valid CEC "wake-up" message,
the code now injects a "KEY_WAKEUP" event in to the Linux UInput
subsystem.  This keyevent is then received by Android's inputflinger
and wakes up the system as if the WAKEUP key had been pressed.

Tested with S0.5, S1 and S2 standby and the system can now wake-up
on receipt on valid CEC "wake-up" messages from the TV.

refs #SWANDROID-3091

Change-Id: I387f96ef6a56d0d05e6daf9a58bca8868a035b25