bomx: Avoid dropping frames pending to display
This is a follow up to 8805f5bf.
There is a possible race condition where the framework
may attempt to re-use an output buffer which hasn't been
displayed by bomx yet. By looking at the serial number of
the buffer sent by the framework and the current frame
pending to be displayed, the bomx decoder is able to avoid
dropping the frame before it is displayed.

refs #SWANDROID-3098

Change-Id: Ia89713eaa5a71297cb4edb1b719af74d2484f723
Signed-off-by: Luis Navarro <navarro@broadcom.com>