SWANDROID-3127: [PM/CEC] STB cannot resume from S2 standby with CEC <Set Stream Path> message
The changes made now ensure that the CEC physical address of the STB
can be read, even if NEXUS is in standby but the CPU is awake.  This
ensures that the CEC physical address is kept up-to-date and valid
for when a TV sends the <Set Stream Path> CEC message to wake-up
the STB from standby.

The CEC API used to obtain the physical address is available even
if NEXUS is in standby.  The hotplug handler ensures that the CEC
physical address is updated, so there is no longer any need to
use NEXUS HDMI output API's to determine the CEC physical address.
This is important, as the HDMI output API's are not available if
NEXUS is in standby, and is the main reason why this bug appeared.

To ellaborate on this, the bug only appeared when in S2 standby,
the STB partially resumed to S0 to service an alarm timer and
returned back to S2 standby.  During this transitional period, if
an HDMI "connected" hot-plug event occurred, it would invalidate
the CEC physical address.  Then if a CEC <Set Stream Path> message
was received from the TV, the HDMI_CEC HAL would try and retrieve
the CEC physical address whilst NEXUS was still in standby and it
would fail.  The result would be that the STB would not determine
this as a valid CEC "wake-up" message and Android would not be
woken up from the sleep state.

refs #SWANDROID-3127

Change-Id: I3b5305f494259285084734f5b8f2c2ead05ffd72