SWANDROID-3225: [PM] Watchdog timer resets STB when in S1 standby
We now stop the Nexus watchdog timer when we enter standby, as it
holds a lock on the module preventing the callback from reliably
firing whilst in standby.  When we resume from standby, we
re-start the Nexus watchdog timer.

Whilst in standby, we continue to pet the kernel watchdog to
prevent the system from rebooting.

refs #SWANDROID-3225

Change-Id: I22ca53af05c71faf8ba05c0eaf832ef307ddd53e