SWANDROID-3228: [PM] Deadlock if app spawned when Nexus is in standby but kernel not suspended
We now ensure that Nexus is in S0 but the display remains off (S0.5 standby)
when the system does not fully suspend (e.g. due to a wake-lock or wakeup source).
This ensures that any Nexus API's are available and allows an app that is
re-spawned to continue without the system deadlocking.

The problem manifests itself when Nexus is told to enter standby (e.g. S2), but
the kernel has not fully suspended due to a wake-lock or wakeup source.  If
an app such as "leanback launcher" happens to die and get re-spawned whilst
Nexus is in standby, then it will deadlock due to Nexus API's like V3D/M2MC
being locked.  This will cause an ANR and Nexus will catastrophically crash.
However, the system will not automatically reboot in this deadlock condition,
as Nexus is in standby and the Nexus watchdog is not available.

refs #SWANDROID-3228

Change-Id: Ibac197237b81a7dc10e2f1fce5aada3acf0686d0