SWANDROID-3090: [PM] Enable WoL support without ethtool (remove SELinux violations)
The Power HAL now checks for the existence of a network interface by
scanning the contents of directory "/sys/class/net" before performing
an ioctl on the interface.  By doing so, we avoid the kernel trying
to insmod a missing network device module and avoid generating an
SELinux sys_module violation.

refs #SWANDROID-3090

Change-Id: I5f946916db8db94931078eff188f8729a9bd50b5