bomx: add a special property to remove any heap toggle.
heap toggle is used to put the nexus picture buffer heaps in
secure or open mode depending on the actual wanted state for
a given decoder mode (secure or open).  this is the replacement
to the former 'urr-toggle' mode.

the property added here allows to prevent any such toggle from
ever occuring which would leave the heaps in the same state as
the boot up state of nexus, which for our current integration based
on the nexus server svp_urr mode means all picture buffer heaps
are secure for-ever.

this mode/property can be useful for debugging but is not intended
to be used as the default behavior since it would cause problems with
open playback and their integration model.

refs #SWANDROID-3270

Change-Id: I97391225b0596c0bcc709e7f8b9f92c39c650db0
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>