bomx: Enable aac secure decoder
The following changes are included in this commit

1. Addition of a new decoder component to handle secure aac.

2. Creation of a shared library (libbomx_secbuff) to manage
the allocation of the secure buffers used for decryption
and decoding. This shared library will also be used by the
crypto plugins to access the secure buffers.

3. Change on the interpratation of the omx secure input buffers.
After this change, the secure decoders (audio and video) will
use memory block handles as input buffers instead of pointers.
The memory block will wrap a new data structure where the pointer
to the secure memory is stored but which may also contain other
data such as a clear memory buffer. A clear memory buffer is
required to transfer the codec config from the crypto plugin to
the decoder in clear memory. The aac decoder requires accessing
the codec config to build adts headers so this data needs to be
in non-restricted memory. The secure video decoder won't use a
clear memory buffer at this point.
It is possible that in the future, this same data structure may
be used to transfer other information from plugin to decoder
when doing svp.

4. A few other cleanup changes.

refs #SWANDROID-1492

Change-Id: I59a6334e5c6619759ca8dd464bb8308105c07df6
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>