Removing nexus logger filesystem entry
Nexus used to require a file to be present in order to route the
server user-space logs into the same process writing out the kernel
logs.  In android this caused some problems and isn't required
since it all goes through the android log service.  This will now
hint to nexus not to use a file mmap and nxlogger will only handle
kernel side logs.

refs #SWANDROID-3320

Change-Id: Ida01f2842df7d4c0e6f8bfe2f8dda46d20589b29