hwc: always flush background on oob-video with multi-gpx composition.
because we cannot actually know the content of the gpx layer when
doing a smart background composition of multiple gpx layers with an
out-of-band video plane underneath, we may actually end up causing the
video to be blocked if the gpx layers net result is fully opaque.

to avoid this visual artifact, we always assume that we need to flush
the background when doing such composition; this adds a fill operation,
but that's only relevant to this particular scenario.

refs #SWANDROID-3262
b/28841976

Change-Id: I2f7bbd2f7567e951c70ab21c24f6b5230deb03e5
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>