[PM] Spurious AON_GPIO wake-up received during shut down
On some platforms (notably the BCM97252SSFFG) when entering S5 shut
down, the BT_DEV_WAKE signal will change state causing the
BT_HOST_WAKE signal to toggle.  Our Power HAL implementation will
react to the AON_GPIO interrupt and will inject a wakeup/power key
event in to the Linux input sub-system.  The Android framework then
receives this wake-up key-event and will then inform the Power HAL to
wake back up ("power_set_interactive()" API called).

This is not desirable, as once Android is told to shut down, then it
should not be interrupted.  Any attempt to try and wake it back up
during the shut down process can cause some parts of Android to
wake-back up, resulting in an unclean shut down.

The simple fix is to still allow the AON_GPIO interrupt(s) to occur,
but not to inject the POWER/WAKEUP key event in to the input
sub-system of Linux.

This fix has been tested with a BCM97252SSFFG platform whilst
placing Android in to S5, and it fixes the problem whereby the
"power_set_interactive()" Power HAL callback was being called whilst
shutting down.  This was tested with Power HAL verbose logging
enabled and the output of logcat was inspected.

refs #SWANDROID-3438

Change-Id: Iab24522ac2ee89f032a2fbf5fe5c9bb40b6b2abf