[PM] Power HAL does not initially set GPIO output
The assumption was made that when Android powers up, it will always
be in power state S0 (i.e. ON).  However, this is not always the
case, as you may want to power-up from state S5 to state S0.5 (where
the display is disconnected).

Because of this wrong assumption, the NexusPower GPIO initialisation
always set the output level to the S0 power state level described for
output GPIO's in the aon_gpio.cfg configuration file.  For BT, this
could mean that the BT_DEV_WAKE signal would be set to an incorrect
level resulting in the BT stack not reacting to WoBLE events.

The fix is to allow the power state level to be provided to the GPIO
initialisation method.  The "initialiseGpios()" and
"uninitialiseGpios()" methods have been decoupled from the NexusPower
class instantiation and destruction methods so that the power state
can be retrieved from the NexusPower class prior to initialising the
GPIO's.

refs #SWANDROID-3439

Change-Id: I7f26922dc5cde22fa5e20e2e905420c2f8d2e9c8