[PM] Random KEY event set for some input GPIO's
The following line shows what a typical GPIO input line looks like for
waking up the STB:

AON_GPIOxx = [Input,RisingEdge,Wakeup]

The following lines show what a typical GPIO input line looks like for not
waking up the STB:

AON_GPIOyy = [Input,FallingEdge,none]
AON_GPIOyy = [Input,FallingEdge]

In the last example, because the third argument is missing there was a bug in
the code that caused the key code to be uninitialised which could result in a
random (invalid) key code being injected in to the Linux input sub-system.

The fix is simply to ensure that the "key" variable in
"NexusPower::NexusGpio::initialise()" is initialised to KEY_RESERVED (i.e. 0).

refs #SWANDROID-3446

Change-Id: I33c43d85b59fcd13b96d68dafc7551ffd44e5551
Signed-off-by: Rob McConnell <robert.mcconnell@broadcom.com>