[PM] Clean-up Power HAL code to avoid race
We now instantiate an event monitoring thread when the Power HAL is
first initialised.  This monitors for shutdown and suspend events
from the droid_pm kernel PM driver and also for interactivity events
from the "set_power_interactive()" Power HAL callback.  This keeps
all event monitoring in one place and avoids the need for constantly
terminating and creating threads when changing power states which
can lead to race conditions.

refs #SWANDROID-3443

Change-Id: I9c69be0efa8be130e6e941195bcbac3f24a1a023
Signed-off-by: Rob McConnell <robert.mcconnell@broadcom.com>