Need to be able to power up Android to S0.5 from S5
Part 1 of 3 changes:
===================

This change allows for Android to be woken up to S0.5 standby from
previously being shutdown (S5).  This permits the behaviour whereby
the STB is placed in to S5, an alarm-timer wake-up event fires and
then the STB is woken back up (cold boot) to S0.5 standby.  This
mechanism allows S5 to be treated almost like an S2 suspended state,
albeit with a long resume time.  This is useful for customers who
require a lower power consumption state than
that provided by S2.

This requirement requires Android to be able to be booted with the A/V
outputs disconnected (display inactive) if a wake-up timer event fires
whilst in S5.  nxserver has been modified to always start with video
outputs disconnected and then enable video outputs if a wake-up timer
event did not occur.  This requirement also means that Android should
enter standby after the framework has booted, if no other wake-up
event other than the wake-up timer occurred.

In order to meet these requirements, we make use of the lid switch
mechanism in the Android framework.  This uses the underlying Linux
input switch facility to provide switch events to user-space
applications.

Android's inputflinger monitors for switch events (including the lid
switch) and passes these to the InputManager Service. The Phone Window
Manager then receives the state of the lid switch and uses this to
decide whether Android should suspend or wake-up.

The Power HAL interfaces with the "sw_lid" droid_pm sysfs attribute in
order for the initial lid state to be set (depending on whether the
STB was woken solely by an alarm timer event or not).  It is also used
to ensure that the state of the lid switch is set to "up" when an
interactive power hint is received.

Because the Android framework always generates an interactive power
hint and calls the "setInteractive" Power HAL callback during boot, we
need to avoid setting the power state (to S0) and sending a lid switch
"up" event, otherwise Android will be brought out of S0.5 standby to
the fully active S0 state.

Part 2 of the changes are for the droid_pm kernel driver.
Part 3 of the changes are for setting an overlay override to enable
the lid switch event to have an effect in the Android framework.

refs #SWANDROID-3373

Change-Id: Icef7068a25d881b0e2cfdb94b496dfe236bef365
Signed-off-by: Rob McConnell <robert.mcconnell@broadcom.com>