[PM] Hang when entering S5 from short press of POWER button
The problem faced was if the STB was configured to enter S5 on a short
press of the POWER button (persist.sys.power.offstate=S5).  On the
press of the POWER button on the remote control, sometimes the STB
would reach the reboot kernel console line, but would fail to be woken
up from this state and would appear to be hung.  Only a full
power-cycle would allow the STB to boot again.

It was discovered that occassionally, the shutdown sequence would
begin prior to the STB receiving the PM resume from suspend callback.
In this situation, the droid_pm PM kernel driver would fail in the
reboot notifier and would return "NOTIFY_BAD".  Upon further
investigation, it was found that the
"wait_for_completion_interruptible_timeout()" in the reboot notifier
would be interrupted with a "-ERESTARTSYS" error code due to the
kernel aborting the suspend sequence (due to the Power HAL
intentionally holding a wake-lock).

The fix ensures that the Power HAL does not begin the actual shutdown
sequence until after the system has begun to suspend, the suspend has
been aborted and the system has resumed again.  This ensures that the
reboot notifier does not get interrupted causing the shutdown sequence
in the kernel to fail.

refs #SWANDROID-3469

Change-Id: I19cca586c17a83a8c0e5b6ea4fa8adeaf5ae9e96
Signed-off-by: Rob McConnell <robert.mcconnell@broadcom.com>