hwc: add ability to punch alpha hole for sideband specifically.
we punch an alpha hole when we detect that the depth of the composition
stack is greater than a threshold and so we can't easily determine or
expect to know whether the hole was punched properly by the application.

in the ideal case where we are below threshold (set to 1 right now -
meaning one or less graphics layer alongside the sideband); we assume
that the graphics layer is properly configured already to punch the hole
itself.

refs #SWANDROID-3428

Change-Id: Ie6e6600cb01de1f1f0bb77eab0c3d099da015db3
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>