mpeg2ts: Set threshold for duration calculation back to 500ms
The duration calculation for mpeg2ts files was not consistent, the
calculation was timing out before finding a duration with a 50ms
consitency. It is not possible to get a reliable 50ms derivation
in under 2 seconds for some files. This change will decrease the accuracy
of the calculation but it makes it consistent between playbacks.

refs #SWANDROID-3476

Change-Id: I220528c3a8107884b0bd2708eeb374ac6650e7d9
Signed-off-by: Derek Smith <dereks@broadcom.com>