[PM] Spurious BT_HOST_WAKE signal received during boot to S0.5
Part 2 of 3 changes:
===================

Part 1 contains the droid_pm kernel driver changes.
Part 2 contains the Power HAL integration changes.
Part 3 contains the device platform changes (e.g. SELinux).

The problem faced is that when transitioning from S5 to S0.5 from an
alarm timer event, the STB can spuradically wake-up to S0 rather than
remain in S0.5 and later suspend.   Investigations proved that the
BT_HOST_WAKE signal was toggling during Android booting, causing the
Power HAL to believe the STB should fully wake-up.  The BT_HOST_WAKE
signal was toggling prematurely and before the BT stack was fully
initialised (before the BT firmware was loaded).

It was decided that a synchronisation mechanism should exist to allow
the BT stack to indicate to the Power HAL when the BT_HOST_WAKE signal
is valid and can be reacted to (i.e. when to spoof a KEY_WAKEUP event).
Investigations proved that a mechanism already exists within the BT
hardware library whereby the BT_WAKE event can be used to know when the
BT stack is awake or asleep.  During boot, the BT_WAKE signal is only
asserted after the BT stack has loaded.

The next piece in the puzzle was how to communicate this timing event to
the Power HAL without making AOSP modifications.  Fortunately, the
designers of the BT hardware library were wise enough to expose this
functionality and the BT hardware library can be configured to write to
"/proc/bluetooth/sleep/bt_write" to indicate whether the BT_WAKE event
is asserted or deasserted.

The changes in part 2 of this commit provide a generic means by which
software modules (e.g. BT stack) can manage when a GPIO input designated
as a "Wakeup" source is enabled.  The Power HAL mow reacts to the
DROID_PM_EVENT_BT_WAKE_ON event to know "when" to begin enabling any and
all input GPIO's tagged with the "BT" attribute in the "aon_gpio.cfg"
configuration file.

To prevent the shutdown acknowledgement from receiving non DROID_PM
resume events and failing (e.g. receiving a DROID_PM_EVENT_BT_WAKE_ON
event whilst suspending/resuming), we now only read events in one
location in the code and that is in the main
"power_event_monitor_thread()".

refs #SWANDROID-3444

Change-Id: Ib082369224b99b8ff0bdc6dbfd029e60fddbabcf