Audio: Corrected the timestamp reported in get_presentation_position
The timestamp reported in get_presentation_position() needs to
correspond to the frame position we report in the same function.
Apparently we just return the current system time which is incorrect.
This applies to audio HAL in all cases (i.e. nexus, direct, and
tunneled).

This commit also protects from an underflow case where bogus data can be
fed into NEXUS.

refs #SWANDROID-3498

Change-Id: I75a5339694131eda913714a5307b471fe27b2b6f