ir: make remote name configurable.
certain game studio (gameloft as example) are canning special names
for controllers that are not keypads.  failure to be recognized
as a controller defaults to keypad mapping which expects to find
keymaps not actually available on a controller.  to work around
this chicken-and-egg problem, we allow the ir controller to be
named anything possible, evidently according to one of the magic
(canned) controller.

we suggest to use "Nexus Remote" as a generic android bound
controller that makes sense; but other controller names are also
canned possibly, just nothing related to broadcom unfortunately, it
is a good idea to pick a controller that name is recognized by most
since we do not switch the name run time (yet?).

http://b/31609319
refs #SWANDROID-3527

Change-Id: I6fef51cbd947ff9ffd5fd0a579f1d016030743d7
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>