bomx/audio: implement accurate seeking with tunneling
Added the following changes:
1) Paused the stc before the tunneling output stream
is closed during seeking. This is to prevent video
to keep running while the new audio output stream
is created.

2) Deferred the audio hal start until the first write
operation is requested. Since audio is driving the stc,
it is better to resume the stc channel until the first
audio sample is available in order to set the stc
properly.

3) Created another stc in host-mode and propagated it
from the audio hal to the video decoder (along the
existing stc used for a/v sync). This new stc is used
to notify the video the resume pts from where the
video frames needs to be displayed after seeking.

4) Force the video decoder to drop the decoded frames
behind the resume pts after seeking.

refs #SWANDROID-3610

Change-Id: Ic3b2f14767321d8d9c56a2404b4dd25c43d17bd1
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>