bomx: Route brcm aac decoder to Android aac default
Only the secure aac decoder requires the Nexus hardware
decoder as it needs access to secure memory.
For non-secure decoding, the default Android software
decoder already provides the required capabilities to
the system. We simply wrap the Android aac decoder
under broadcom vendor's name in order to comply with
several gts testcases which expect the root name of the
secure/non-secure decoders to match.

refs #SWANDROID-3358

Change-Id: I14cc20045813038f9ce91eaab4abe3e0b1c4d27d
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>