bomx: another improvement to tunnel seeking
When the video decoder is in the process of dropping
frames to reach the start pts, we need to pause the
stc or else audio will continue playing and dragging
the stc forward. This will end up causing video
artifacts in the form of fast-forward video playback
for frames with pts between the start-pts and the stc.
This is particularly important for high fps streams
since it takes longer for the video decoder to
decode and drop frames before reaching the start pts.

refs #SWANDROID-3610

Change-Id: I57b4efaab981a1e8890d31568cfa677a44fb741b
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>