media: do not request decoder(s) not supported by hardware.
'legacy' decoders are: vp8, h263, mpeg4; attempting to request those
on a platform that does not offer them (even when mangled with valid
decoders) will cause nexus to fail matching any valid decoder regardless
of the final decoder intent due to the unicity of the decoder hardware.

this is not the most elegant, but certainly the least invasive, so
will do for now.

refs #SWANDROID-3767

Change-Id: Ibda9454afdb94f9ac8aae153d50aba0ae5328ebe
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>