audio: additional tunneling seeking updates
1) The stc handle will always be created and
owned by the audio device and not by the output
stream. This keeps the implementation more simple
and also avoids mistmatch errors between the stc
used by the audio HAL and the one used by the bomx
video decoder. When seeking, the video decoder
instance is preserved but the output stream may
be re-created and therefore there's the potential
of stc mismatch when the stc handle is owned by
the output stream.

2) Make sure that the stc handle is not 'frozen'
when the tunneling output stream is closed in order
to allow the seeking operation to work properly.

3) Initializing the sync-stc to 0xFFFFFFFF instead
of zero, otherwise seeking to position zero would
not work.

4) A couple minor corrections in the partial timestamp
header logic.

refs #SWANDROID-3610

Change-Id: Idec8f31e6fd3548d9e291cdb322ec02d27100145
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>