tif: include BcmHdmiTvInput in the build
The HDMI TV input service is now fixed and can be safely included in
the build. The service will not run by default, because the TV input
HAL checks the property ro.nx.trim.hdmiin and doesn't notify the
service if it's equal to 1 (default).

Note: if you want to use the HDMI TV input, you also need to change the
kernel memory configuration, not just the property.

refs #SWANDROID-3728

Change-Id: I49ecefc290e8bd560d17ec507d9a4f1269d6a215