nxserver: cleanup wdog processing and integrate initial (bootloader) wdog.
if the bootloader started the wdog; nxserver must terminate it before taking
over control (which includes the case of no watchdog wanted).

the nxserver takes control of the watchdog support at the time android issues
the sys.bootcomplete event; prior to this event, the watchdog may have been
set by bootloader and so may run up to the point of handover.

when running watchdog from bootloader, one needs to be aware that a first time
boot (or a factory reset boot) will take more time than any subsequent boot and
that therefore the bootloader timeout must account for this extended duration
to avoid false trigger prior to reaching the sys.bootcomplete state (since the
sys.bootcomplete event will be issued at a later time in those cases, due to initial
processing such as: userdata partition encryption, dexopt run); as example a 10
minutes (600 seconds) bootloader timeout is a good value.

refs #SWANDROID-3848

Change-Id: I8474d8b62019108cfbbd560dcb66f0700c4838d0
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>