hwc2: split out fb in "android-fb" and "platform-fb".
define two distinct fb which are directed to different comsumer:

- "android-fb" is the fb target that android is aware of and which
  is used to determine the user interface; changing android-fb may
  also require changing the dpi and aapt resources configuration to
  best match the gui.
- "platform-fb" is the fb known to the middleware and consumed by it;
  hwc2 in particular takes inputs from "android-fb" content layers and
  push them into "platform-fb".

in most cases, "android-fb" == "platform-fb"; this model is a pass
through.  however it is now possible to make "android-fb" != "platform-fb"
in which case a scaling will take place via the m2mc at the time of
composition of the android-fb elements into the platform-fb. scaling in
m2mc can be an alternative to scaling in gfd (bypass client) or nsc.

refs #SWANDROID-3870

Change-Id: I703ea6df863569258ec05b93af58e56d5721c31f
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>