splash: Fix custom splash and Netflix startup reason
Decouple the logic between launching from cold boot and launching from resume.  The
property dyn.nx.boot.wakeup was overloaded and caused confusion between cold boot and
resume.  On cold boot, the only condition requiring a custom splash is when
PARTNER_CUSTOMIZATION is received, we show custom splash and launch Netflix if the box
was powered on by the Netflix key.  This knowledge is passed from NexusIrInput::start()
to the BcmCustomizer in dyn.nx.boot.key2.  On resume, we show custom splash and launch
Netflix if Netflix key was pressed and the box just woke up from resume.  The property
dyn.nx.boot.wakeup is used to indicate the wake up condition and checked in the
KeyEvent handler of BcmCustomizer.

Also fixed the content of the power_on extra field in NETFLIX_KEY intent to reflect
actual launch reason.

refs #SWANDROID-3973

Change-Id: I3e41fe48f2421a904ec259469aeb7f40fdc81215
Signed-off-by: Edwin Chan <edwin.chan@broadcom.com>