sideband viewer: clean up.
Fixed the overlay banner.
Don't attempt to lock the canvas and draw in the sideband surface. The
surface can only have one producer, which is the sideband content.
Protect sideband context and player with mutex.

Sync-up the file player with the corresponding nxclient play.c example.
Put the player inside a thread, so we can use the same loop as play.c
(easier for side-by-side diff)

refs #SWANDROID-4003

Change-Id: Ic01dd4eb54979a83f652049715a31b6abc7fffd4