BOMX: Add timeout to smooth format change
In the case when the app does not release all the output buffers back to
the frameworks during a resolution change, our decoder can be stuck in a
state waiting for certain output buffers before it notifies the
frameworks about the format change event. When this scenario happens,
adding a timeout to stop waiting and notify the frameworks about the
format change event.

refs #SWANDROID-3826

Change-Id: Idae9366a9237d39ee4f9e6dde819993063e7b75a