bomx: Adjust buffer return rate for tunneling
The logic added in swandroid-3624 to speed-up the
return of input buffers causes side effects on
tunneling playback. The high rate of buffers
return results in too much data consumed by the
playpump. This indirectly causes the framework
to accummulate too many 'Output Rendered' events,
eventually causing an assert when the number of
events reaches 64 (the capacity of AMessage).
This fix simply disables the high-speed return
logic for tunneling playback.

refs #SWANDROID-3941

Change-Id: I045911e7a3c372b716ff20cff187f92f65c1511b
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>